# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: EDU Power Push Backend API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    API for campus electricity scanning, product user accounts, personal and
    campus statistics, push notifications, and the operator admin panel.
    Processes bind to loopback by default (Compose maps only 127.0.0.1). Public
    access is same-origin reverse proxy + TLS; operators may also use an SSH
    tunnel. Campus aggregates (`/api/v1/campus/*`) are anonymous; personal data
    uses short-lived access JWTs; admin and raw meter endpoints require an
    operator/admin session or `ADMIN_TOKEN` (never ship that token in a browser build).
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/audit:
    get:
      tags: [Admin]
      operationId: listAdminAudit
      summary: Recent panel writes
      description: Credential changes record which fields changed, never their values.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/users:
    get:
      tags: [Admin]
      operationId: listAdminUsers
      summary: Account list for the panel
      description: The q parameter matches email, nickname and meter number at once.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string, enum: [user, operator, admin] }
        - name: status
          in: query
          schema: { type: string, enum: [active, disabled] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/admin/users/{user_id}:
    patch:
      tags: [Admin]
      operationId: updateAdminUser
      summary: Change a role or enable/disable an account
      description: >
        Requires admin, because whoever can grant access is effectively an administrator.
        Disabling also revokes that account's sessions. Demoting or disabling the last active
        admin is rejected, as is changing your own role or status.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [user, operator, admin]
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Would lock everyone out (last_admin) or lock yourself out (self_lockout)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Admin]
      operationId: deleteAdminUser
      summary: Permanently delete an account
      description: >
        Requires admin. Deletes account-owned bindings, sessions, notification settings,
        channel credentials and push logs through cascading foreign keys. Meter readings,
        official daily usage and bills remain intact. Deleting yourself or the last active
        administrator is rejected.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Would delete the caller or the last active administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/users/{user_id}/meter:
    delete:
      tags: [Admin]
      operationId: unbindAdminUserMeter
      summary: Clear an account's active meter binding
      description: >
        Requires admin. This only closes the current user-meter binding. Historical meter
        readings, official daily usage and bills are retained.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Binding cleared, including when it was already empty
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/users/{user_id}/sessions/revoke:
    post:
      tags: [Admin]
      operationId: revokeAdminUserSessions
      summary: Sign an account out everywhere
      description: Gentler than disabling, because the user can simply sign in again.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Number of revoked sessions
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Internal administrator token from ADMIN_TOKEN.
    userAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived product-user access JWT.
    refreshCookie:
      type: apiKey
      in: cookie
      name: edu_power_refresh
      description: Long-lived rotating refresh JWT stored as an HttpOnly SameSite cookie.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid administrator token, access JWT, or refresh JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated operation is disabled by product policy
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    CaptchaProvider:
      type: string
      enum: [disabled, turnstile]
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        provider:
          $ref: "#/components/schemas/CaptchaProvider"
        challenge_url:
          type: string
          format: uri-reference
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    AdminUser:
      type: object
      required:
        [id, email, nickname, role, status, created_at, active_sessions, place, channels_enabled]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        nickname: { type: string }
        role:
          type: string
          enum: [user, operator, admin]
        status:
          type: string
          enum: [active, disabled]
        created_at: { type: string, format: date-time }
        last_login_at: { type: [string, "null"], format: date-time }
        email_verified_at:
          type: [string, "null"]
          format: date-time
          description: Set when registration or an email change verifies ownership.
        active_sessions:
          type: integer
          description: Unrevoked, unexpired refresh sessions, i.e. roughly the number of devices.
        meter: { type: [string, "null"] }
        place: { type: string }
        channels_enabled: { type: integer }
    AuditEntry:
      type: object
      required: [id, actor_label, action, target, detail, request_id, created_at]
      properties:
        id: { type: string, format: uuid }
        actor_label: { type: string }
        action:
          type: string
          example: settings.mail.save
        target: { type: string }
        detail:
          type: object
          additionalProperties: true
          description: Metadata only. Credential values are never recorded here.
        request_id: { type: string }
        created_at: { type: string, format: date-time }
...